split in up colorbar method:
- rename Plotly.Colorbar -> Colorbar.draw
